+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # SkyHanni - Change Log
 
-## Version 0.16.UNRELEASED (UNRELEASED)
+## Version 0.15.1 (2023-01-25)
 
 ## Features
 + Adding green line around items that are clickable. (Inside the **Not Clickable Items Feature**)